requirements.txt
flask==2.3.3
requests==2.31.0
paystack-python==2.0.0
python-dotenv==1.0.0
pytest==7.4.0
supabase==2.3.5
psycopg2-binary==2.9.9
